+++ Proyecto final/Pixel Picture/src/main/java/backend/pixel_picture/Repositorios/EncargoRepository.java
@@ -3,5 +3,5 @@ package backend.pixel_picture.Repositorios;
 import backend.pixel_picture.Entidades.EncargoFotografia;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EncargoRepository extends JpaRepository<EncargoFotografia, Integer> {
+public interface EncargoRepository extends JpaRepository<EncargoFotografia, Long> {
 }
